requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-dotenv
httpx
PyMuPDF
python-docx
pandas
openpyxl
Pillow
jinja2
groq
reportlab
google-generativeai
